Migrate Query.operationsStats
